Use global raytracer for lua scene
This way we can have command line options.
Would love to instead have LuaScene as a class and then pass it in a
RayTracer object, but we would need to use some external library to bind
methods to lua functions. That is annoying (may make this not work in
the lab) so we will not do it.