Uniform grid now working
Not optimal, since it is using bounding boxes to determine which cells
an object goes into.  This could be made a lot faster if meshes were
broken up into a list of faces (polygons), and the faces were individually
placed into cells.

In order to do that we would probably have to use a lot more memory, or
add some weird new Primitive like "Polygon", which would probably be a
horrible hack.