Rework materials idea
Now there is a base material which is a PhongMaterial.  This material
does the lighting calculations. It delegates a kd() function to get the
diffuse colour given a hitRecord (i.e. point of intersection).

The idea is that a TextureMaterial (or FunctionMaterial) can override
the getKd() method to look up what colour that particular point should
be.